# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MT-Engine",
    "description": "M-Team \u514d\u8d39\u79cd\u5b50\u730e\u624b",
    "version": "6.6.2"
  },
  "paths": {
    "/api/pilot/config": {
      "get": {
        "summary": "Get Config",
        "description": "Get current pilot configuration\n\nReturns:\n    dict: Current configuration",
        "operationId": "get_config_api_pilot_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update Config",
        "description": "Update pilot configuration\n\nArgs:\n    config: New configuration\n\nReturns:\n    dict: Status message",
        "operationId": "update_config_api_pilot_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/pilot/dry-run": {
      "get": {
        "summary": "Dry Run",
        "description": "Simulate run - returns what would be downloaded/cleaned\n\nReturns:\n    dict: Download candidates and cleanup candidates",
        "operationId": "dry_run_api_pilot_dry_run_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/pilot/run-download": {
      "post": {
        "summary": "Trigger Download",
        "description": "Manually trigger download cycle\n\nReturns:\n    dict: Status message",
        "operationId": "trigger_download_api_pilot_run_download_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/pilot/run-cleanup": {
      "post": {
        "summary": "Trigger Cleanup",
        "description": "Manually trigger cleanup cycle\n\nReturns:\n    dict: Status message",
        "operationId": "trigger_cleanup_api_pilot_run_cleanup_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/torrents": {
      "get": {
        "summary": "Get Torrents",
        "description": "\u83b7\u53d6\u79cd\u5b50\u5217\u8868",
        "operationId": "get_torrents_api_torrents_get",
        "parameters": [
          {
            "name": "discount",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Discount"
            }
          },
          {
            "name": "min_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Size"
            }
          },
          {
            "name": "max_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Size"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutomationConfig": {
        "properties": {
          "download": {
            "$ref": "#/components/schemas/DownloadPolicy"
          },
          "cleanup": {
            "$ref": "#/components/schemas/CleanupPolicy"
          },
          "enable_notification": {
            "type": "boolean",
            "title": "Enable Notification",
            "default": true
          }
        },
        "type": "object",
        "title": "AutomationConfig",
        "description": "Main pilot configuration (credentials via env vars, not stored)"
      },
...
      "CleanupPolicy": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "min_share_ratio": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Share Ratio",
            "default": 0.0
          },
          "min_seed_time_hours": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min Seed Time Hours",
            "default": 1
          },
          "max_download_time_hours": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Download Time Hours",
            "default": 12
          },
          "dead_seed_minutes": {
            "type": "integer",
            "minimum": 5.0,
            "title": "Dead Seed Minutes",
            "default": 30
          },
          "dead_seed_max_ratio": {
            "type": "number",
            "minimum": 0.0,
            "title": "Dead Seed Max Ratio",
            "default": 0.01
          },
          "min_current_users": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min Current Users",
            "default": 5
          },
          "min_upload_speed_kbps": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min Upload Speed Kbps",
            "default": 200
          },
          "elimination_ratio": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Elimination Ratio",
            "default": 0
          }
        },
        "type": "object",
        "title": "CleanupPolicy",
        "description": "Cleanup policy configuration"
      },
...
      "DownloadPolicy": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "max_active_tasks": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Active Tasks",
            "default": 20
          },
          "interval_seconds": {
            "type": "integer",
            "minimum": 60.0,
            "title": "Interval Seconds",
            "default": 300
          },
          "save_path": {
            "type": "string",
            "title": "Save Path",
            "default": "/downloads/mt_free_farm"
          },
          "disk_usage_threshold": {
            "type": "integer",
            "maximum": 95.0,
            "minimum": 50.0,
            "title": "Disk Usage Threshold",
            "default": 90
          },
          "rules": {
            "$ref": "#/components/schemas/RuleConfig"
          }
        },
        "type": "object",
        "title": "DownloadPolicy",
        "description": "Download policy configuration"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuleConfig": {
        "properties": {
          "min_size_gb": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Size Gb",
            "default": 20.0
          },
          "max_size_gb": {
            "type": "number",
            "minimum": 0.0,
            "title": "Max Size Gb",
            "default": 500.0
          },
          "discount_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Discount Types",
            "default": [
              "FREE",
              "_2X_FREE"
            ]
          },
          "include_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Include Keywords",
            "default": []
          },
          "exclude_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclude Keywords",
            "default": [
              "AUDIOBOOK"
            ]
          },
          "max_seeders": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Seeders",
            "default": 10
          },
          "min_leechers": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min Leechers",
            "default": 100
          },
          "weight_size": {
            "type": "number",
            "maximum": 10.0,
            "minimum": -10.0,
            "title": "Weight Size",
            "default": -1.0
          },
          "weight_free_time": {
            "type": "number",
            "maximum": 10.0,
            "minimum": -10.0,
            "title": "Weight Free Time",
            "default": 0.4
          },
          "weight_age": {
            "type": "number",
            "maximum": 10.0,
            "minimum": -10.0,
            "title": "Weight Age",
            "default": 0.5
          },
          "weight_seeders": {
            "type": "number",
            "maximum": 10.0,
            "minimum": -10.0,
            "title": "Weight Seeders",
            "default": 3.0
          }
        },
        "type": "object",
        "title": "RuleConfig",
        "description": "Rule configuration for filtering and scoring torrents"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "MT Engine API key sent as an Authorization bearer token.",
        "scheme": "bearer"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "MT Engine API key sent in the X-MT-ENGINE-Key header.",
        "in": "header",
        "name": "X-MT-ENGINE-Key"
      }
    }
  }
}